requirements.txt
pdfplumber
PyPDF2
pandas
scikit-learn
streamlit
pillow
imagehash
sentence-transformers
torch
